Move the tests around
