+++ src/Camelotia.Tests/Camelotia.Tests.csproj
