+++ src/Camelotia.Tests/Presentation/AuthViewModelTests.cs
@@ -8,7 +8,7 @@ using NSubstitute;
 using ReactiveUI;
 using Xunit;
 
-namespace Camelotia.Presentation.Tests.ViewModels
+namespace Camelotia.Tests.Presentation
 {
     public sealed class AuthViewModelTests
     {
