+++ src/Camelotia.Tests/Presentation/HostAuthViewModelTests.cs
@@ -8,33 +8,38 @@ using NSubstitute;
 using ReactiveUI;
 using Xunit;
 
-namespace Camelotia.Presentation.Tests.ViewModels
+namespace Camelotia.Tests.Presentation
 {
-    public sealed class DirectAuthViewModelTests
+    public sealed class HostAuthViewModelTests
     {
         private readonly IProvider _provider = Substitute.For<IProvider>();
 
         [Fact]
         public void LoginCommandShouldStayDisabledUntilInputIsValid()
         {
-            var model = BuildDirectAuthViewModel();
+            var model = BuildHostAuthViewModel();
             model.Login.CanExecute(null).Should().BeFalse();
+            model.Address = "10.10.10.10";
             model.Username = "hello";
             model.Password = "world";
+            model.Port = "5000";
             model.Login.CanExecute(null).Should().BeTrue();
         }
 
         [Fact]
         public void HasErrorsShouldTriggerWhenProviderBreaks()
         {
-            _provider.DirectAuth("hello", "world").Returns(x => throw new Exception("example"));
-            
-            var model = BuildDirectAuthViewModel();
+            _provider.HostAuth("10.10.10.10", 5000, "hello", "world").Returns(x => throw new Exception("example"));
+                
+            var model = BuildHostAuthViewModel();
             model.HasErrors.Should().BeFalse();
                 
+            model.Port = "5000";
             model.Username = "hello";
             model.Password = "world";
+            model.Address = "10.10.10.10";
             model.Login.Execute(null);
+            
             model.HasErrors.Should().BeTrue();
             model.ErrorMessage.Should().Be("example");
         }
@@ -42,22 +47,43 @@ namespace Camelotia.Presentation.Tests.ViewModels
         [Fact]
         public void ShouldBeBusyWhenLoggingIn()
         {
-            _provider.DirectAuth("hello", "world").Returns(new Task(() => { }));
+            _provider.HostAuth("10.10.10.10", 5000, "hello", "world").Returns(new Task(() => { }));
             
-            var model = BuildDirectAuthViewModel();
+            var model = BuildHostAuthViewModel();
             model.IsBusy.Should().BeFalse();
-            
+
+            model.Port = "5000";
             model.Username = "hello";
             model.Password = "world";
+            model.Address = "10.10.10.10";
             model.Login.Execute(null);
             model.IsBusy.Should().BeTrue();
         }
 
-        private DirectAuthViewModel BuildDirectAuthViewModel()
+        [Fact]
+        public void ShouldTreatNonIntegersAsInvalid()
+        {
+            var model = BuildHostAuthViewModel();
+            model.Login.CanExecute(null).Should().BeFalse();
+            
+            model.Port = "5000";
+            model.Username = "hello";
+            model.Password = "world";
+            model.Address = "10.10.10.10";
+            model.Login.CanExecute(null).Should().BeTrue();
+
+            model.Port = "abc";
+            model.Login.CanExecute(null).Should().BeFalse();
+            
+            model.Port = "42";
+            model.Login.CanExecute(null).Should().BeTrue();
+        }
+
+        private HostAuthViewModel BuildHostAuthViewModel()
         {
             RxApp.MainThreadScheduler = Scheduler.Immediate;
             RxApp.TaskpoolScheduler = Scheduler.Immediate;
-            return new DirectAuthViewModel(_provider);
+            return new HostAuthViewModel(_provider);
         }
     }
 }
