+++ src/Camelotia.Tests/Presentation/MainViewModelTests.cs
@@ -11,7 +11,7 @@ using NSubstitute;
 using ReactiveUI;
 using Xunit;
 
-namespace Camelotia.Presentation.Tests.ViewModels
+namespace Camelotia.Tests.Presentation
 {
     public sealed class MainViewModelTests
     {
