+++ src/Camelotia.Tests/Presentation/OAuthViewModelTests.cs
@@ -8,7 +8,7 @@ using NSubstitute;
 using ReactiveUI;
 using Xunit;
 
-namespace Camelotia.Presentation.Tests.ViewModels
+namespace Camelotia.Tests.Presentation
 {
     public sealed class OAuthViewModelTests
     {
