+++ src/Camelotia.Tests/Presentation/ProviderViewModelTests.cs
@@ -12,7 +12,7 @@ using NSubstitute;
 using ReactiveUI;
 using Xunit;
 
-namespace Camelotia.Presentation.Tests.ViewModels
+namespace Camelotia.Tests.Presentation
 {
     public sealed class ProviderViewModelTests
     {
