+++ src/Camelotia.Tests/Services/VkontakteProviderTests.cs
@@ -1,4 +1,5 @@
-using System;
+﻿using System;
+using System.IO;
 using Akavache;
 using Camelotia.Services.Models;
 using Camelotia.Services.Providers;
@@ -6,27 +7,27 @@ using FluentAssertions;
 using NSubstitute;
 using Xunit;
 
-namespace Camelotia.Presentation.Tests.Services
+namespace Camelotia.Tests.Services
 {
-    public sealed class GithubProviderTests
+    public sealed class VkontakteFileSystemProviderTests
     {
         private readonly IBlobCache _blobCache = Substitute.For<IBlobCache>();
         private readonly ProviderModel _model = new ProviderModel
         {
             Id = Guid.NewGuid(),
-            Created = DateTime.Now,
-            Type = "GitHub"
+            Type = "Vkontakte",
+            Created = DateTime.Now
         };
 
         [Fact]
         public void VerifyDefaultPropertyValues()
         {
-            var provider = new GitHubProvider(_model, _blobCache);
-            provider.InitialPath.Should().Be(string.Empty);
+            var provider = new VkDocsProvider(_model, _blobCache);
+            provider.InitialPath.Should().Be(Path.DirectorySeparatorChar.ToString());
 
             provider.CanCreateFolder.Should().BeFalse();
             provider.Created.Should().Be(_model.Created);
-            provider.Name.Should().Be("GitHub");
+            provider.Name.Should().Be("Vkontakte");
             provider.Id.Should().Be(_model.Id);
 
             provider.SupportsDirectAuth.Should().BeTrue();
